Move project template tests into create transfer test file
