+++ end-to-end-tests/cypress/fixtures/transfersLinks.json
@@ -1,17 +1,42 @@
 [
     "home",
-    "cookie-preferences", 
-    "transfers/trustname",
     "cookie-preferences",
+    "transfers/newtransfersinformation",
+    "transfers/trustname",
     "transfers/trustsearch?query=company&change=False",
-    "project/10002562",
-    "project-assignment/10002562",
-    "project/10002562/features",
-    "project/10002562/transfer-dates",
-    "project/10002562/legalrequirements",
-    "project/10002562/rationale",
-    "project/10002562/benefits",
-    "project/10002562/academy-and-trust-information",
-    "project/10002562/advisory-board/preview?",
-    "project/10002562/advisory-board/download?" 
+    "transfers/outgoingtrustdetails?query=SOUTH+YORK+MULTI+ACADEMY+TRUST&change=False&trustId=10067112",
+    "transfers/incomingtrust",
+    "transfers/searchincomingtrust?query=bishop+fraser&change=False",
+    "project/10003257",
+    "project-assignment/10003257",
+    "project/10003257/features",
+    "project/10003257/features/reason",
+    "project/10003257/features/type",
+    "project/10003257/transfer-dates",
+    "project/10003257/transfer-dates/advisory-board-date",
+    "project/10003257/transfer-dates/target-date",
+    "project/10003257/benefits",
+    "project/10003257/benefits/intended-benefits",
+    "project/10003257/benefits/risks",
+    "project/10003257/benefits/equalities-impact-assessment",
+    "project/10003257/legalrequirements",
+    "project/10003257/legalrequirements/outgoing-trust-consent",
+    "project/10003257/legalrequirements/incoming-trust-agreement",
+    "project/10003257/legalrequirements/diocesan-consent",
+    "project/10003257/rationale",
+    "project/10003257/rationale/project",
+    "project/10003257/rationale/trust-or-sponsor",
+    "project/10003257/academy-and-trust-information",
+    "project/10003257/academy-and-trust-information/recommendation",
+    "project/10003257/school-data/10064723",
+    "project/10003257/general-information/10064723",
+    "project/10003257/pupil-number/10064723",
+    "project/10003257/latest-ofsted-judgement/10064723",
+    "project/10003257/latest-ofsted-judgement/10064723?addOrEditAdditionalInformation=true#additional-information-hint",
+    "project/10003257/key-stage-4-performance/10064723",
+    "project/10003257/key-stage-4-performance/10064723?addOrEditAdditionalInformation=true#additional-information-hint",
+    "project/10003257/key-stage-5-performance/10064723",
+    "project/10003257/key-stage-5-performance/10064723?addOrEditAdditionalInformation=true#additional-information-hint",
+    "project/10003257/advisory-board/preview?",
+    "project/10003257/advisory-board/download?" 
   ]
