+++ end-to-end-tests/cypress/pages/download.ts
@@ -2,11 +2,9 @@ class DownloadPage {
 
   public slug = 'advisory-board/download'
 
-  public downloadProjectTemplate(projectId): this {
+  public downloadProjectTemplate(): this {
 
     cy.get('h1').should('contain.text', 'Download project template')
-
-    cy.get('[data-test="download-htb"]').should('contain.text', `SAT-${projectId}`)
 
     cy.get('[data-test="download-htb"]').click()
